factored some code into a new module for bootstrap updated bootstrap updated characters in application manifest added availability to template
