+++ app/views/apps_medium.jade
@@ -10,6 +10,7 @@ block content
       h4(class="listing-heading") Watch Video
 
     div(class="medium-video")
+      p available? #{medium.available}
       video(
         id="medium-video",
         src="http://data.goddard.com/media/#{medium.uri}",
