+++ test/fixtures/apps.json
@@ -186,7 +186,7 @@
           "source": "http://globalhealthmedia.org/portfolio-items/attaching-your-baby-at-the-breast/?portfolioID=5623",
           "available": false,
           "size": 107358050,
-          "uri": "gem/breastfeeding/Attaching%20Your%20Baby%20at%20the%20Breast%20%e2%80%93%20Breastfeeding%20Series-wjt-Ashodw8.mp4",
+          "uri": "gem/breastfeeding/Attaching%20Your%20Baby%20at%20the%20Breast%20-%20Breastfeeding%20Series-wjt-Ashodw8.mp4",
           "description": "Good attachment will help a baby get more milk and make breastfeeding more comfortable. This video shows why good attachment is so important to breastfeeding success and what a mother can do to deeply attach her baby to her breast."
         },
         {
@@ -194,7 +194,7 @@
           "source": "http://globalhealthmedia.org/portfolio-items/breastfeeding-in-the-first-hours-after-birth/?portfolioID=5623",
           "available": false,
           "size": 92390145,
-          "uri": "gem/breastfeeding/Breastfeeding%20in%20the%20First%20Hours%20After%20Birth%20%e2%80%93%20Breastfeeding%20Series-uMcgJR8ESRc.mp4",
+          "uri": "gem/breastfeeding/Breastfeeding%20in%20the%20First%20Hours%20After%20Birth%20-%20Breastfeeding%20Series-uMcgJR8ESRc.mp4",
           "description": "Within the first few hours after birth, most healthy newborns will instinctively move to their mother’s breast and attach on their own. This video shows early breastfeeding initiation through the journeys of 3 newborns and ways to ensure the success of this vital practice for any new mother and baby."
         },
         {
@@ -218,7 +218,7 @@
           "available": false,
           "size": 63797149,
           "source": "http://globalhealthmedia.org/portfolio-items/how-to-express-breastmilk/?portfolioID=5623",
-          "uri": "gem/breastfeeding/How%20to%20Express%20Breastmilk%20%e2%80%93%20Breastfeeding%20Series-axQi5PqRZ0M.mkv",
+          "uri": "gem/breastfeeding/How%20to%20Express%20Breastmilk%20-%20Breastfeeding%20Series-axQi5PqRZ0M.mkv",
           "description": "Expressing breastmilk can relieve breast fullness, soften the breasts, and make it easier for the milk to flow. It is also important to maintain a mother’s milk supply during separation and make milk available for someone else to feed the baby. This video shows how to hand express breastmilk."
         },
         {
@@ -226,7 +226,7 @@
           "available": false,
           "size": 67392150,
           "source": "http://globalhealthmedia.org/portfolio-items/increasing-your-milk-supply/?portfolioID=5623",
-          "uri": "gem/breastfeeding/Increasing%20Your%20Milk%20Supply%20%e2%80%93%20Breastfeeding%20Series-M_NLHiasZok.mp4",
+          "uri": "gem/breastfeeding/Increasing%20Your%20Milk%20Supply%20-%20Breastfeeding%20Series-M_NLHiasZok.mp4",
           "description": "Most mothers are able to provide their babies with all the milk they need. However, if a baby is not thriving as she should, this video shows a mother steps to improve her milk supply and the baby’s feeding."
         },
         {
@@ -234,7 +234,7 @@
           "available": false,
           "size": 50654569,
           "source": "http://globalhealthmedia.org/portfolio-items/is-your-baby-getting-enough-milk/?portfolioID=5623",
-          "uri": "gem/breastfeeding/Is%20Your%20Baby%20Getting%20Enough%20Milk%20%e2%80%93%20Breastfeeding%20Series-LX1fl8NtTuw.mp4",
+          "uri": "gem/breastfeeding/Is%20Your%20Baby%20Getting%20Enough%20Milk%20-%20Breastfeeding%20Series-LX1fl8NtTuw.mp4",
           "description": "A common concern of a new mother is whether her baby is getting enough milk. This video shows signs to tell whether the baby is feeding well."
         },
         {
@@ -242,7 +242,7 @@
           "available": false,
           "size": 81286937,
           "source": "http://globalhealthmedia.org/portfolio-items/what-to-do-about-breast-pain/?portfolioID=5623",
-          "uri": "gem/breastfeeding/What%20To%20Do%20About%20Breast%20Pain%20%e2%80%93%20Breastfeeding%20Series-XpGRuRAWqKM.mp4",
+          "uri": "gem/breastfeeding/What%20To%20Do%20About%20Breast%20Pain%20-%20Breastfeeding%20Series-XpGRuRAWqKM.mp4",
           "description": "Breast pain in a breastfeeding mother is commonly caused by one of three problems: engorgement, blocked ducts, or breast infections. This video shows how a mother can know which problem she has and what she can do to help herself."
         },
         {
@@ -250,7 +250,7 @@
           "available": false,
           "size": 58987807,
           "source": "http://globalhealthmedia.org/portfolio-items/what-to-do-about-nipple-pain/?portfolioID=5623",
-          "uri": "gem/breastfeeding/What%20To%20Do%20About%20Nipple%20Pain%20%e2%80%93%20Breastfeeding%20Series-WVEABNhXr1A.mp4",
+          "uri": "gem/breastfeeding/What%20To%20Do%20About%20Nipple%20Pain%20-%20Breastfeeding%20Series-WVEABNhXr1A.mp4",
           "description": "Nipple pain in a breastfeeding mother is common but not normal. It often happens when mothers first start breastfeeding and is usually due to a baby not attaching deeply. This video shows how nipple pain happens and what a mother can do to help herself."
         }
       ],
@@ -357,7 +357,7 @@
               "source": "http://globalhealthmedia.org/portfolio-items/positions-for-breastfeeding/?portfolioID=5623",
               "available": false,
               "size": 86958983,
-              "uri": "gem/newborn_and_child_birth/Positions%20for%20Breastfeeding%20%e2%80%93%20Breastfeeding%20Series-y9kKp3YnNYY.mp4",
+              "uri": "gem/newborn_and_child_birth/Positions%20for%20Breastfeeding%20-%20Breastfeeding%20Series-y9kKp3YnNYY.mp4",
               "description": "The way a mother holds her baby affects how easy it is for her baby to feed effectively. This video shows basic points important for any position and then demonstrates several common positions that have worked well for many mothers and babies."
             },
             {
